refactor: consolidate usStates module
